feat(replaceMedia): add skipCheck to continue without existing media entry

+++ src/scripts/replaceMedia.js
@@ -11,7 +11,7 @@ import { uploadMedia } from 'graphql/mutations/CreateMediaArticle';
 /**
  * @param {object} args
  */
-async function replaceMedia({ articleId, url } = {}) {
+async function replaceMedia({ articleId, url, skipCheck = false } = {}) {
   const {
     body: { _source: article },
   } = await client.get({ index: 'articles', type: 'doc', id: articleId });
@@ -22,7 +22,7 @@ async function replaceMedia({ articleId, url } = {}) {
   const oldMediaEntry = await mediaManager.get(article.attachmentHash);
 
   /* istanbul ignore if */
-  if (!oldMediaEntry)
+  if (!skipCheck && !oldMediaEntry)
     throw new Error(
       `Article ${articleId}'s attachment hash "${
         article.attachmentHash
@@ -35,9 +35,8 @@ async function replaceMedia({ articleId, url } = {}) {
   });
 
   console.info(
-    `Article ${articleId} attachment hash: ${oldMediaEntry.id} --> ${
-      newMediaEntry.id
-    }`
+    `Article ${articleId} attachment hash: ${oldMediaEntry?.id ??
+      article.attachmentHash} --> ${newMediaEntry.id}`
   );
   await client.update({
     index: 'articles',
@@ -50,16 +49,18 @@ async function replaceMedia({ articleId, url } = {}) {
     },
   });
 
-  return Promise.all(
-    oldMediaEntry.variants.map(variant =>
-      oldMediaEntry
-        .getFile(variant)
-        .delete()
-        .then(() => {
-          console.info(`Old media entry variant=${variant} deleted`);
-        })
-    )
-  );
+  if (oldMediaEntry) {
+    return Promise.all(
+      oldMediaEntry.variants.map(variant =>
+        oldMediaEntry
+          .getFile(variant)
+          .delete()
+          .then(() => {
+            console.info(`Old media entry variant=${variant} deleted`);
+          })
+      )
+    );
+  }
 }
 
 export default replaceMedia;
@@ -79,6 +80,10 @@ if (require.main === module) {
         description: 'The URL to the content to replace',
         type: 'string',
         demandOption: true,
+      },
+      skipCheck: {
+        description: 'Skip old media entry check',
+        type: 'boolean',
       },
     })
     .help('help').argv;
